move RSXformProp
