+++ package/cpp/rnskia/dom/nodes/JsiAtlasNode.h
@@ -3,6 +3,7 @@
 #include "JsiDomDrawingNode.h"
 #include "RectProp.h"
 #include "SkImageProps.h"
+#include "RSXformProp.h"
 
 #include <memory>
 #include <tuple>
